added valid flag to C++ ClientProxy and fixed a bunch of compiler warnings

+++ client_libs/c++/commsproxy.cc
@@ -50,7 +50,7 @@ CommsProxy::CommsProxy(PlayerClient* pc, unsigned short index, unsigned char acc
 // Destructor
 CommsProxy::~CommsProxy()
 {
-  for(int i=0;i<this->msg_num;i++)
+  for(size_t i=0;i<this->msg_num;i++)
     free(this->msg[i]);
   free(this->msg);
   free(this->msg_len);
@@ -107,11 +107,11 @@ void CommsProxy::FillData(player_msghdr_t hdr, const char* buffer)
 // Delete a message (and shift the rest down)
 int CommsProxy::Delete(int index)
 {
-  if((index < 0) || (index >= this->msg_num))
+  if((index < 0) || (index >= (int)this->msg_num))
     return(-1);
 
   free(this->msg[index]);
-  for(int i=index;i<this->msg_num-1;i++)
+  for(size_t i=index;i<this->msg_num-1;i++)
   {
     this->msg[i] = this->msg[i+1];
     this->msg_len[i] = this->msg_len[i+1];
@@ -130,10 +130,10 @@ void CommsProxy::Print()
 {
   printf("# Comms(%d:%d) - %c : %d messages\n", device, index, access,
          this->msg_num);
-  for(int i=0;i<this->msg_num;i++)
+  for(size_t i=0;i<this->msg_num;i++)
   {
     printf("# len %d msg [%s]\n", this->msg_len[i], this->msg[i]);
-    printf("# timestamp: %d:%d\n", 
+    printf("# timestamp: %ld:%ld\n", 
            this->msg_ts[i].tv_sec, this->msg_ts[i].tv_usec);
   }
   return;
